feat: #31, #32, #65: Add argument validation to Dashboard endpoint. Update tests.

+++ mcm-backend/mcm-core/src/main/java/at/ac/tuwien/model/change/management/core/service/DashboardServiceImpl.java
@@ -1,9 +1,14 @@
 package at.ac.tuwien.model.change.management.core.service;
 
+import at.ac.tuwien.model.change.management.core.exception.DashboardNotFoundException;
 import at.ac.tuwien.model.change.management.core.mapper.neo4j.DashboardEntityMapper;
 import at.ac.tuwien.model.change.management.core.model.Dashboard;
 import at.ac.tuwien.model.change.management.core.model.UserRole;
+import at.ac.tuwien.model.change.management.core.service.validator.DashboardValidator;
 import at.ac.tuwien.model.change.management.graphdb.dao.DashboardEntityDAO;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -18,19 +23,24 @@ public class DashboardServiceImpl implements DashboardService {
 
     private final DashboardEntityDAO repository;
     private final DashboardEntityMapper mapper;
+    private final DashboardValidator validator;
 
     @Override
-    public Dashboard createDashboard(Dashboard dashboard) {
+    public Dashboard createDashboard(@NotNull Dashboard dashboard) {
+        validator.validateForCreate(dashboard);
         return mapper.fromEntity(repository.save(mapper.toEntity(dashboard)));
     }
 
     @Override
-    public Dashboard getDashboard(String id) {
-        return mapper.fromEntity(repository.getById(id));
+    public Dashboard getDashboard(@NotBlank String id) {
+        var found = repository.getById(id);
+        if (found == null)
+            throw new DashboardNotFoundException("No dashboard exists with id " + id);
+        return mapper.fromEntity(found);
     }
 
     @Override
-    public List<Dashboard> getDashboardsForRoles(List<UserRole> roles) {
+    public List<Dashboard> getDashboardsForRoles(@NotEmpty List<UserRole> roles) {
 
         //Query all dashboards that contain at least one of the user roles.
 
@@ -45,7 +55,10 @@ public class DashboardServiceImpl implements DashboardService {
     }
 
     @Override
-    public void deleteDashboard(String id) {
+    public void deleteDashboard(@NotBlank String id) {
+        var found = repository.getById(id);
+        if (found == null)
+            throw new DashboardNotFoundException("No dashboard exists with id " + id);
         repository.deleteById(id);
     }
 }
